+++ backend/routes/purchase-order.js
@@ -245,7 +245,7 @@ router.post('/', async (req, res) => {
     // Compute total GST from items
     const totalGst = Array.isArray(items)
       ? items.reduce((sum, item) => {
-          const lineTotal = (parseFloat(item.unit_price) || 0) * (parseInt(item.quantity) || 0);
+          const lineTotal = (parseFloat(item.unit_price) || 0) * (parseFloat(item.quantity) || 0);
           const rate      = parseFloat(item.gst_rate) || 0;
           return sum + (lineTotal * rate) / 100;
         }, 0)
@@ -288,7 +288,7 @@ router.post('/', async (req, res) => {
           [
             poDbId,
             itemDbId,
-            parseInt(item.quantity)    || 1,
+            parseFloat(item.quantity)  || 1,
             parseFloat(item.unit_price) || 0,
             item.gst_type  || null,
             parseFloat(item.gst_rate)  || 0,
@@ -343,7 +343,7 @@ router.put('/:po_no', async (req, res) => {
     // Compute total GST from items
     const totalGst = Array.isArray(items)
       ? items.reduce((sum, item) => {
-          const lineTotal = (parseFloat(item.unit_price) || 0) * (parseInt(item.quantity) || 0);
+          const lineTotal = (parseFloat(item.unit_price) || 0) * (parseFloat(item.quantity) || 0);
           const rate      = parseFloat(item.gst_rate) || 0;
           return sum + (lineTotal * rate) / 100;
         }, 0)
@@ -386,7 +386,7 @@ router.put('/:po_no', async (req, res) => {
           [
             poDbId,
             itemDbId,
-            parseInt(item.quantity)    || 1,
+            parseFloat(item.quantity)  || 1,
             parseFloat(item.unit_price) || 0,
             item.gst_type  || null,
             parseFloat(item.gst_rate)  || 0,
