+++ backend/routes/quotation.js
@@ -162,7 +162,7 @@ router.post('/', async (req, res) => {
         const itemDbId = itemRes.rows[0].id;
         await client.query(
           'INSERT INTO quotation_items (quotation_id, item_id, quantity, unit_price, unit, company_id) VALUES ($1, $2, $3, $4, $5, $6)',
-          [quotationDbId, itemDbId, parseInt(item.quantity) || 1, parseFloat(item.unit_price) || 0, item.unit || 'Piece', req.user.company_id]
+          [quotationDbId, itemDbId, parseFloat(item.quantity) || 1, parseFloat(item.unit_price) || 0, item.unit || 'Piece', req.user.company_id]
         );
       }
     }
@@ -238,7 +238,7 @@ router.put('/:quotation_no', async (req, res) => {
         const itemDbId = itemRes.rows[0].id;
         await client.query(
           'INSERT INTO quotation_items (quotation_id, item_id, quantity, unit_price, unit, company_id) VALUES ($1, $2, $3, $4, $5, $6)',
-          [quotationDbId, itemDbId, parseInt(item.quantity) || 1, parseFloat(item.unit_price) || 0, item.unit || 'Piece', req.user.company_id]
+          [quotationDbId, itemDbId, parseFloat(item.quantity) || 1, parseFloat(item.unit_price) || 0, item.unit || 'Piece', req.user.company_id]
         );
       }
     }
