+++ deskManager/src/pages/panel/PoPanel.jsx
@@ -194,7 +194,7 @@ export default function PoPanel({ purchaseOrder, quotation, processRq, tradeId, 
 
   // Calculated totals
   const itemsBasic = (purchaseOrder.items || []).reduce(
-    (s, i) => s + (parseFloat(i.unit_price) || 0) * (parseInt(i.quantity) || 0), 0
+    (s, i) => s + (parseFloat(i.unit_price) || 0) * (parseFloat(i.quantity) || 0), 0
   );
   const gstTotal   = parseFloat(purchaseOrder.gst)             || 0;
   const transport  = parseFloat(purchaseOrder.transport)       || 0;
@@ -269,7 +269,7 @@ export default function PoPanel({ purchaseOrder, quotation, processRq, tradeId, 
               </thead>
               <tbody className="divide-y divide-slate-100 bg-white">
                 {(purchaseOrder.items || []).map((item, idx) => {
-                  const lineBasic = (parseFloat(item.unit_price) || 0) * (parseInt(item.quantity) || 0);
+                  const lineBasic = (parseFloat(item.unit_price) || 0) * (parseFloat(item.quantity) || 0);
                   const lineGst   = lineBasic * ((parseFloat(item.gst_rate) || 0) / 100);
                   const lineTotal = lineBasic + lineGst;
                   return (
@@ -283,7 +283,7 @@ export default function PoPanel({ purchaseOrder, quotation, processRq, tradeId, 
                       <td className="px-4 py-3 text-slate-700 font-medium max-w-[160px] truncate">{item.description || '—'}</td>
                       <td className="px-4 py-3 text-right font-bold text-slate-800">{item.quantity}</td>
                       <td className="px-4 py-3 text-right font-bold text-slate-800">
-                        {Math.max(0, (parseInt(item.quantity) || 0) - (parseInt(item.delivered_qty) || 0))}
+                        {Math.max(0, (parseFloat(item.quantity) || 0) - (parseFloat(item.delivered_qty) || 0))}
                       </td>
                       <td className="px-4 py-3 text-right font-mono font-bold text-slate-800">₹{fmt(item.unit_price)}</td>
                       <td className="px-4 py-3 text-right">
