+++ deskManager/src/pages/panel/QuotationPanel.jsx
@@ -54,7 +54,7 @@ export default function QuotationPanel({ quotation, rfq, tradeId, tradeType, onR
   };
 
   const itemsTotal = (quotation.items || []).reduce(
-    (a, i) => a + (parseFloat(i.unit_price) || 0) * (parseInt(i.quantity) || 0), 0
+    (a, i) => a + (parseFloat(i.unit_price) || 0) * (parseFloat(i.quantity) || 0), 0
   );
 
   return (
@@ -119,7 +119,7 @@ export default function QuotationPanel({ quotation, rfq, tradeId, tradeType, onR
               </thead>
               <tbody className="divide-y divide-slate-100 bg-white">
                 {(quotation.items || []).map((item, idx) => {
-                  const lineTotal = (parseFloat(item.unit_price) || 0) * (parseInt(item.quantity) || 0);
+                  const lineTotal = (parseFloat(item.unit_price) || 0) * (parseFloat(item.quantity) || 0);
                   return (
                     <tr key={idx} className="hover:bg-slate-50 transition-colors">
                       <td className="px-4 py-3">
